Define kie-maven-plugin version explicitly in all examples
Due to the non-standard build we have (many SNAPSHOT deps between
different repos), we can not just configure the plugin in parent
(like before the change) as that prevents us from bulding the examples.
What happens is that the plugin version is set to a version which is not
yet available (it will only be available after the
droolsjbpm-integration repo is built), and since the versions are
updated as a first step, the plugin is not available at that time.

Using the ${version.org.kie} is all the modules is sligthly
inconvenient, but it also means that the plugin will actually stay in
the POM, even after being processed by the XSLT. This is a good thing
as the POM is otherwise invalid (it uses unknown packaging kjar).